reorg token

+++ src/ByteCodeGenJob.h
@@ -25,7 +25,7 @@ struct AstArrayPointerSlicing;
 struct TypeInfoPointer;
 struct DataSegment;
 enum class ByteCodeOp : uint16_t;
-enum class TokenId : uint16_t;
+enum class TokenId : uint8_t;
 
 static const uint32_t BCC_FLAG_NOLOCATION = 0x00000001;
 static const uint32_t BCC_FLAG_NOSAFETY   = 0x00000002;
